+++ src/components/index.ts
@@ -16,4 +16,6 @@ export * from './Grid'
 
 export * from './Switch'
 
+export * from './Input'
+
 export * from './Skelton'