+++ types/components/Input/index.stories.d.ts
@@ -1,0 +1,9 @@
+import { Input } from '.';
+import type { Meta, StoryObj } from '@storybook/react';
+declare const meta: Meta<typeof Input>;
+export default meta;
+type Story = StoryObj<typeof Input>;
+export declare const Default: Story;
+export declare const Close: Story;
+export declare const Moving: Story;
+//# sourceMappingURL=index.stories.d.ts.map