+++ types/libs/animation/variant/type.d.ts
@@ -1,4 +1,5 @@
 import type { CSSProperties } from 'react';
+import type { AnimationProps } from './AnimationFactory';
 export type AnimationBaseProps = {
     delay?: CSSProperties['animationDelay'];
     duration?: CSSProperties['animationDuration'];
@@ -8,5 +9,8 @@ export type AnimationBaseProps = {
 export type GeneratedAnimationValue = {
     className: string;
     style: object;
+};
+export type AnimationArgs = {
+    animationProps?: AnimationProps;
 };
 //# sourceMappingURL=type.d.ts.map