+++ types/src/components/CheckBox/index.css.d.ts
@@ -1,6 +1,0 @@
-export declare const checkContainerStyle: string;
-export declare const inputDefaultStyle: string;
-export declare const iconStyle: string;
-export declare const containerStyle: string;
-export declare const labelStyle: string;
-//# sourceMappingURL=index.css.d.ts.map