+++ types/src/index.d.ts
@@ -1,3 +1,0 @@
-/** @format */
-export * from './components';
-//# sourceMappingURL=index.d.ts.map