+++ types/src/libs/animation/genAnimation.d.ts
@@ -1,6 +1,0 @@
-import type { AnimationProps } from './variant/AnimationFactory';
-export declare const resolveAnimation: (props: AnimationProps) => {
-    className: string;
-    style: object;
-};
-//# sourceMappingURL=genAnimation.d.ts.map