+++ types/src/libs/animation/variant/Squeeze/index.d.ts
@@ -1,3 +1,0 @@
-export type { SqueezeProps } from './type';
-export { genAnimation as genSqueezeAnimation } from './gen';
-//# sourceMappingURL=index.d.ts.map