+++ types/src/libs/animation/variant/Squeeze/type.d.ts
@@ -1,9 +1,0 @@
-import type { AnimationBaseProps } from '../type';
-export type OptionProps = {
-    axis?: 'x' | 'y';
-};
-export type SqueezeProps = {
-    key: 'squeeze';
-    option: OptionProps & AnimationBaseProps;
-};
-//# sourceMappingURL=type.d.ts.map