+++ types/src/libs/animation/variant/TextColor/index.d.ts
@@ -1,3 +1,0 @@
-export type { TextColorProps } from './type';
-export { genAnimation as genTextColorAnimation } from './gen';
-//# sourceMappingURL=index.d.ts.map